Add sinon and chai for testing
We decided we'll go with using es6 import syntax for sinon and chai tools. This is really nice because the tools are build to work that way out of the box, so there's actually no configuration needed to get this to work. We just have to install the libraries, import them into our spec files, and angular cli/webpack does the rest.